⌛Refactor: 오류 수정 및 Resolution 에 사용자 식별 로직 추가

+++ src/main/kotlin/com/example/days/domain/resolution/service/ResolutionServiceImpl.kt
@@ -4,18 +4,23 @@ import com.example.days.domain.resolution.dto.request.ResolutionRequest
 import com.example.days.domain.resolution.dto.response.ResolutionResponse
 import com.example.days.domain.resolution.model.Resolution
 import com.example.days.domain.resolution.repository.ResolutionRepository
+import com.example.days.domain.user.repository.UserRepository
+import com.example.days.global.infra.security.AuthenticationUtil.getAuthenticationUserId
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class ResolutionServiceImpl(
-    private val resolutionRepository: ResolutionRepository
+    private val resolutionRepository: ResolutionRepository,
+    private val userRepository: UserRepository
 ): ResolutionService {
     @Transactional
     override fun createResolution(request: ResolutionRequest): ResolutionResponse {
-        // TODO : 사용자 식별 가능한 로직 추가 예졍
-        val resolution = resolutionRepository.save(ResolutionRequest.of(request))
+        val user = getAuthenticationUserId()
+            .let { userRepository.findByIdOrNull(it) }
+            ?: TODO("예외처리")
+        val resolution = resolutionRepository.save(ResolutionRequest.of(request, user))
         return ResolutionResponse.from(resolution)
     }
 
@@ -30,17 +35,24 @@ class ResolutionServiceImpl(
 
     @Transactional
     override fun updateResolution(resolutionId: Long, request: ResolutionRequest): ResolutionResponse {
-        // TODO : 목표 작성자만 수정 가능하도록 제한
+        val user = getAuthenticationUserId()
         val updatedResolution = getByIdOrNull(resolutionId)
-        updatedResolution.updateResolution(request.title, request.description, request.category)
-        return ResolutionResponse.from(updatedResolution)
+        if(updatedResolution.author.id == user){
+            updatedResolution.updateResolution(request.title, request.description, request.category)
+            return ResolutionResponse.from(updatedResolution)
+        }
+        else TODO("예외처리")
+
     }
 
     @Transactional
     override fun deleteResolution(resolutionId: Long) {
-        // TODO : 목표 작성자만 삭제 가능하도록 제한
+        val user = getAuthenticationUserId()
         val resolution = getByIdOrNull(resolutionId)
-        resolutionRepository.delete(resolution)
+        if (resolution.author.id == user){
+            resolutionRepository.delete(resolution)
+        }
+        else TODO("예외처리")
     }
 
     fun getByIdOrNull(id: Long) = resolutionRepository.findByIdOrNull(id) ?: TODO("예외처리 구현예정")
